+++ srcs/Libft/srcs/DispTxt/ft_puthex.c
@@ -6,14 +6,14 @@
 /*   By: siferrar <siferrar@student.le-101.fr>      +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/10/10 20:42:44 by shimon       #+#   ##    ##    #+#       */
-/*   Updated: 2020/01/07 18:20:52 by siferrar    ###    #+. /#+    ###.fr     */
+/*   Updated: 2020/01/09 15:37:39 by siferrar    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
 #include "../../includes/libft.h"
 
-size_t	ft_puthex(uintptr_t n, int upper)
+size_t	ft_puthex(unsigned long n, int upper)
 {
 	size_t count;
